+++ example-app/router/index.js
@@ -7,13 +7,14 @@ module.exports = function (app) {
   app.use('/api',
     router.use('/', require('./api')),
     router.use('/auth', require('./auth')),
-    router.use('/saml', require('./saml')),
+    router.use('/mongo-demo', require('./mongo-demo')),
+    router.use('/nexmo', require('./nexmo')),
     router.use('/oidc', require('./oidc')),
     router.use('/oauth', require('./oauth')),
-    router.use('/mongo-demo', require('./mongo-demo')),
+    router.use('/saml', require('./saml')),
+    router.use('/sse', require('./sse')),
     router.use('/t4t', require('./t4t')),
     router.use('/webpush', require('./webpush')),
-    router.use('/nexmo', require('./nexmo'))
   )
 
   const { KNEXFILE } = global.CONFIG