+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 - [chore] update packages & cleanup & work on improving documentation
 - [example-app] add server side event (SSE) example
 - [example-app] add serve-index function to uploads folder (list files and folders)
+- [example-app] multiple UPLOAD_STATIC & UPLOAD_MEMORY
 - In Progress
-  - [example-app] multiple UPLOAD_STATIC
-  - [feat] add commitizen
+  - [chore] remove example-webpack
+  - [chore] remove example-app/routes/ related to example-webpack (check transaction code)
+  - [feat] add commitizen, add semantic-version
   - [example-app] jsonschema & openapi 3.1
     - shared webworkers (https://blog.logrocket.com/real-time-processing-web-workers/)
     - nodejs workers